+++ clients/admin-ui/src/features/datastore-connections/ConnectionMenu.tsx
@@ -10,12 +10,25 @@ import {
 import React from "react";
 
 import { MoreIcon } from "../common/Icon";
+import DeleteConnectionModal from "./DeleteConnectionModal";
+import DisableConnectionModal from "./DisableConnectionModal";
+import { AccessLevel, ConnectionType } from "./types";
 
 interface ConnectionMenuProps {
-  // disabled: boolean;
+  connection_key: string;
+  disabled: boolean;
+  name: string;
+  connection_type: ConnectionType;
+  access_type: AccessLevel;
 }
 
-const ConnectionMenu: React.FC<ConnectionMenuProps> = () => (
+const ConnectionMenu: React.FC<ConnectionMenuProps> = ({
+  connection_key,
+  disabled,
+  connection_type,
+  access_type,
+  name,
+}) => (
   <Menu>
     <MenuButton
       as={Button}
@@ -33,18 +46,14 @@ const ConnectionMenu: React.FC<ConnectionMenuProps> = () => (
         >
           <Text fontSize="sm">Edit</Text>
         </MenuItem>
-        <MenuItem
-          _focus={{ color: "complimentary.500", bg: "gray.100" }}
-          // onClick={handleViewDetails}
-        >
-          <Text fontSize="sm">Disable</Text>
-        </MenuItem>
-        <MenuItem
-          _focus={{ color: "complimentary.500", bg: "gray.100" }}
-          // onClick={handleViewDetails}
-        >
-          <Text fontSize="sm">Delete</Text>
-        </MenuItem>
+        <DisableConnectionModal
+          connection_key={connection_key}
+          disabled={disabled}
+          connection_type={connection_type}
+          access_type={access_type}
+          name={name}
+        />
+        <DeleteConnectionModal connection_key={connection_key} />
       </MenuList>
     </Portal>
   </Menu>